make --help output a bit more narrow
The three parameters

  * --wait-event=
  * --wait-event-and-download=
  * --capture-tethered=

all accept an optional argument

    COUNT, SECONDS, MILLISECONDS or MATCHSTRING

which makes the left column of the "gphoto2 --help" output
so wide that the descriptions move *very* far to the right.

Even on a 100 char wide terminal, the "gphoto2 --help" output
is very badly readable due to the very narrow space available
in the right column for the option descriptions.

This replaces the argument variable with "EVENT", and then
adds an epilog line in the "gphoto2 --help" output which
explains that EVENT can be either COUNT, SECONDS, MILLISECONDS
or MATCHSTRING.

This makes the left column narrow enough that the right column
with the description becomes readable again.

This changes the set of translated strings, but no changes to
po/*.po files are included in this commit, due to many unrelated
changes overdue in po/*.po.