Fix mtime part of test001, test002
Commit 0ae34dd has added mtime to the output of gphoto2 -L,
but failed to adapt the two test cases involving gphoto2 -L,
namely test001 and test002.

As the file mtime varies depending on the time of the git
clone or the last git pull or similarly uncontrollable events,
the one reasonable thing to do here is to ignore the actual
value of the mtime and just verify that there *is* an mtime.

Fixes: https://github.com/gphoto/gphoto2/issues/453