Use consistently named non-reserved CPP macros for header files
C preprocessor macros beginning/ending with underscores are
reserved for system use as well as macros containing a double
underscore anywhere, so application software macros are
supposed to be named differently.

For a header file gp-params.h, this consistently uses a
macro name GPHOTO2_GP_PARAMS_H for the usual

            #ifndef GPHOTO2_GP_PARAMS_H
            #define GPHOTO2_GP_PARAMS_H
            ... content of the header file ...
            #endif /* !defined(GPHOTO2_GP_PARAMS_H) */

Found using "clang -Weverything".

This should work, unless when run with an ancient C preprocessor
which might only the first 8 or 10 characters to distinguish
macro names and ignores the remainder of the name.