Remove unused and unsafe system(3) based --hook-script implementation
Back in 2006, I wrote a safe execve(3) based alternative implementation
and we have been using that ever since.

I left the old unsafe system(3) based implementation in the source
file #ifdef'ed out just in case someone needed it, but that does not
appear to have happened, so we can get rid of it.